Move the grizzly adapter into base class (to be used by GitHubPluginTest too)
