+++ src/test/groovy/se/diabol/scrolls/JiraAdapter.groovy
@@ -2,34 +2,16 @@ package se.diabol.scrolls
 
 import com.sun.grizzly.tcp.Request
 import groovy.json.JsonBuilder
-import com.sun.grizzly.tcp.Response
-import com.sun.grizzly.util.buf.ByteChunk
-import com.sun.grizzly.tcp.Adapter
 
 /**
  * Grizzly Adapter that simulates jira REST api for unit test purpose
  */
-class JiraAdapter implements Adapter {
+class JiraAdapter extends UnitTestAdapter {
 
     def projects = JiraData.projects
     def issues = JiraData.issues
 
-    public void service(Request request, Response response) {
-        println "JiraAdapter received request '${request.unparsedURI()}'"
-        def content = new ByteChunk()
-        request.doRead(content)
-        def bytes = null
-        def processResult = processRequest(request)
-        bytes = processResult.content.bytes
-        response.status = processResult.status
-        def chunk = new ByteChunk()
-        chunk.append(bytes, 0, bytes.length)
-        response.contentLength = bytes.length
-        response.contentType = 'text/json'
-        response.outputBuffer.doWrite(chunk, response)
-        response.finish()
-    }
-
+    @Override
     def processRequest(Request request) {
         def result = [:]
         if (!request.getHeader('Authorization').startsWith('Basic')) {
@@ -49,11 +31,4 @@ class JiraAdapter implements Adapter {
         }
         return result
     }
-
-    public void afterService(Request request, Response response) {
-        request.recycle()
-        response.recycle()
-    }
-
-    public void fireAdapterEvent(string, object) {}
 }
